Moved CourseChoicesBaseController
