+++ app/controllers/candidate_interface/continuous_applications/course_choices/course_study_mode_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class CourseStudyModeController < BaseController
+      class CourseStudyModeController < CandidateInterface::CourseChoices::BaseController
         include CandidateInterface::CourseChoices::Concerns::DuplicateCourseRedirect
 
       private
