+++ app/controllers/candidate_interface/continuous_applications/course_choices/do_you_know_which_course_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class DoYouKnowWhichCourseController < BaseController
+      class DoYouKnowWhichCourseController < CandidateInterface::CourseChoices::BaseController
       private
 
         def step_params
