+++ app/controllers/candidate_interface/continuous_applications/course_choices/duplicate_course_selection_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class DuplicateCourseSelectionController < BaseController
+      class DuplicateCourseSelectionController < CandidateInterface::CourseChoices::BaseController
         before_action :set_course
         before_action :set_backlink, only: [:new] # rubocop:disable Rails/LexicallyScopedActionFilter
         skip_before_action :redirect_to_your_applications_if_maximum_amount_of_choices_have_been_used
