+++ app/controllers/candidate_interface/continuous_applications/course_choices/provider_selection_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class ProviderSelectionController < BaseController
+      class ProviderSelectionController < CandidateInterface::CourseChoices::BaseController
       private
 
         def step_params
