+++ app/controllers/candidate_interface/continuous_applications/course_choices/reached_reapplication_limit_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class ReachedReapplicationLimitController < BaseController
+      class ReachedReapplicationLimitController < CandidateInterface::CourseChoices::BaseController
         before_action :set_course
         skip_before_action :redirect_to_your_applications_if_maximum_amount_of_choices_have_been_used
