+++ app/controllers/candidate_interface/continuous_applications/course_choices/review_and_submit_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class ReviewAndSubmitController < BaseController
+      class ReviewAndSubmitController < CandidateInterface::CourseChoices::BaseController
         before_action SubmissionPermissionFilter
         before_action :redirect_to_your_applications_if_submitted
         before_action :redirect_to_course_choice_review_unless_ready_to_submit
