+++ app/controllers/candidate_interface/continuous_applications/course_choices/review_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class ReviewController < BaseController
+      class ReviewController < CandidateInterface::CourseChoices::BaseController
         def show
           @application_choice = current_application.application_choices.find(params[:application_choice_id])
         end
