+++ app/controllers/candidate_interface/continuous_applications/course_choices/review_interruption_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class ReviewInterruptionController < BaseController
+      class ReviewInterruptionController < CandidateInterface::CourseChoices::BaseController
         before_action :redirect_to_your_applications_if_submitted
 
         def show
