+++ app/controllers/candidate_interface/continuous_applications/course_choices/which_course_are_you_applying_to_controller.rb
@@ -1,7 +1,7 @@
 module CandidateInterface
   module ContinuousApplications
     module CourseChoices
-      class WhichCourseAreYouApplyingToController < BaseController
+      class WhichCourseAreYouApplyingToController < CandidateInterface::CourseChoices::BaseController
       private
 
         def current_step
